updated to seperate functionality of model and database
