+++ HaulageApplication/Haulage/DatabaseExecutionServices/VehicleExecutionService.cs
@@ -1,0 +1,34 @@
+﻿using SQLite;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Haulage.BaseClasses;
+using Haulage.BaseClasses.Accounting;
+
+namespace Haulage.DatabaseExecutionServices
+{
+    public static class VehicleExecutionService
+    {
+        public static List<Vehicle> GetVehicles() 
+        {
+            var vehicles = new List<Vehicle>();
+            var sql = "SELECT [VehicleId]" +
+                ",[tripID]    " +
+                ",[LicensePlate]     " +
+                ",[Capacity]      " +
+                ",[DriverId]  " +
+                ",[Status]  " +
+                "FROM [Vehicle];";
+
+            using (var connection = new SQLiteConnection(DatabaseSetup.GetDatabasePath()))
+            {
+                var command = new SQLite.SQLiteCommand(connection);
+                command.CommandText = sql;
+                vehicles = command.ExecuteQuery<Vehicle>();
+            }
+            return vehicles;
+        }
+    }
+}
